add more blacksmithing code, move item classes
