+++ IslandMUD/IslandMUD.vcxproj
@@ -113,6 +113,9 @@
     <ClInclude Include="game.h" />
     <ClInclude Include="generator.h" />
     <ClInclude Include="item.h" />
+    <ClInclude Include="item\equipment.h" />
+    <ClInclude Include="Item\item.h" />
+    <ClInclude Include="item\materials.h" />
     <ClInclude Include="message.h" />
     <ClInclude Include="non_player_character.h" />
     <ClInclude Include="npc_enemy.h" />
